+++ src/utils/collate.py
@@ -22,17 +22,17 @@ def collate_fn(dataset_items: list[dict]) -> dict[Union[torch.Tensor, list]]:
             ],
             batch_first=True,
             padding_value=-1,
-        ),
+        ).long(),
         "user_id": torch.tensor(
             [int(item["user_id"]) for item in dataset_items],
-            dtype=torch.int32,
+            dtype=torch.long,
         ),
     }
 
     if "holdout" in dataset_items[0].keys():
         result_batch["holdout"] = torch.tensor(
             [int(item["holdout"]) for item in dataset_items],
-            dtype=torch.int32,
+            dtype=torch.long,
         )
 
     return result_batch
